+++ webpack.base.config.js
@@ -10,6 +10,10 @@ module.exports = {
       'laxar-mocks': './laxar-mocks.js'
    },
    module: {
+      noParse: [
+         /node_modules[/]ajv[/]dist[/]ajv.bundle.js/,
+         /bower_components\/angular*\/*.js/
+      ],
       loaders: [
          {
             test: /\.js$/,